# packages/shared/openapi/openapi.generated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Varasaan API",
    "version": "0.2.0",
    "description": "Contract-first backend API for Varasaan."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/v1/payments/checkout": {
      "post": {
        "tags": [
          "payments"
        ],
        "operationId": "createCheckout",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentCheckoutRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentCheckoutResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/payments/webhook": {
      "post": {
        "tags": [
          "payments"
        ],
        "operationId": "paymentWebhook",
        "parameters": [
          {
            "name": "X-Razorpay-Signature",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentWebhookRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentWebhookResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/payments/{orderId}": {
      "get": {
        "tags": [
          "payments"
        ],
        "operationId": "getPayment",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentStatusResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/trusted-contacts": {
      "post": {
        "tags": [
          "trusted-contacts"
        ],
        "operationId": "createTrustedContact",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrustedContactCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contact created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrustedContactResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "trusted-contacts"
        ],
        "operationId": "listTrustedContacts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Trusted contacts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrustedContactResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trusted-contacts/invite/accept": {
      "post": {
        "tags": [
          "trusted-contacts"
        ],
        "operationId": "acceptTrustedContactInvite",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invite accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiMessage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/api/v1/trusted-contacts/{trustedContactId}": {
      "delete": {
        "tags": [
          "trusted-contacts"
        ],
        "operationId": "revokeTrustedContact",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TrustedContactId"
          }
        ],
        "responses": {
          "200": {
            "description": "Revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiMessage"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/trusted-contacts/{trustedContactId}/invite": {
      "post": {
        "tags": [
          "trusted-contacts"
        ],
        "operationId": "inviteTrustedContact",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TrustedContactId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrustedContactInviteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invite queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrustedContactInviteResponse"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
...
      "TrustedContactId": {
        "name": "trustedContactId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication failed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "TooManyRequests": {
        "description": "Rate limited",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    },
    "schemas": {
      "ApiMessage": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
...
      "PaymentCheckoutRequest": {
        "type": "object",
        "required": [
          "amount_paise"
        ],
        "properties": {
          "amount_paise": {
            "type": "integer",
            "minimum": 1
          },
          "currency": {
            "type": "string",
            "default": "INR"
          }
        }
      },
      "PaymentCheckoutResponse": {
        "type": "object",
        "required": [
          "order_id",
          "provider",
          "provider_order_id",
          "amount_paise",
          "currency",
          "status"
        ],
        "properties": {
          "order_id": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "provider_order_id": {
            "type": "string"
          },
          "checkout_key_id": {
            "type": "string",
            "nullable": true
          },
          "amount_paise": {
            "type": "integer"
          },
          "currency": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "created",
              "authorized",
              "captured",
              "failed",
              "refunded"
            ]
          }
        }
      },
      "PaymentWebhookRequest": {
        "type": "object",
        "required": [
          "order_id",
          "status",
          "event_sequence"
        ],
        "properties": {
          "event_id": {
            "type": "string",
            "nullable": true
          },
          "order_id": {
            "type": "string"
          },
          "payment_id": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string"
          },
          "event_sequence": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "PaymentWebhookResponse": {
        "type": "object",
        "required": [
          "accepted",
          "processed",
          "reason"
        ],
        "properties": {
          "accepted": {
            "type": "boolean"
          },
          "processed": {
            "type": "boolean"
          },
          "reason": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "PaymentStatusResponse": {
        "type": "object",
        "required": [
          "order_id",
          "status",
          "event_sequence"
        ],
        "properties": {
          "order_id": {
            "type": "string"
          },
          "payment_id": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "created",
              "authorized",
              "captured",
              "failed",
              "refunded"
            ]
          },
          "event_sequence": {
            "type": "integer"
          }
        }
      },
...
      "TrustedContactCreateRequest": {
        "type": "object",
        "required": [
          "name",
          "email",
          "role"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "enum": [
              "viewer",
              "packet_access",
              "recovery_assist"
            ]
          },
          "recovery_enabled": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "TrustedContactInviteRequest": {
        "type": "object",
        "properties": {
          "force_reissue": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "TrustedContactResponse": {
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "role",
          "status"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "enum": [
              "viewer",
              "packet_access",
              "recovery_assist"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "active",
              "revoked"
            ]
          }
        }
      },
...
      "TrustedContactInviteResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "invite_token": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  }
}